requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
pandas>=2.2.0
tenacity==8.2.3

# Web scraping and automation
beautifulsoup4==4.12.2
playwright==1.40.0
greenlet==3.0.1
lxml==5.1.0
tldextract==5.1.1

# Google Custom Search API
google-api-python-client==2.111.0

# AI/LLM (optional - add when implementing AI features)
# openai==1.6.1
# anthropic==0.8.1
google-genai==0.6.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
rich==13.7.0

# Tests
pytest==8.3.3
